refactoring to better deal with 2D and 3D sgrids
